Fix code quality issues for CodeScene gates
- Refactor backendConfig.js into smaller modules (envDetector, urlUtils, envResolver, validator)
- Fix broken markdown links in public/index.md
- Remove unused React imports from new components
- Remove unused catch parameters and variables
- Fix unescaped apostrophes in ContributorGuide.js
- Remove unused Info import from ContributorGuide.js
- Remove unused eslint-disable directives

This addresses:
- New code health below 10.00 (backendConfig.js refactoring)
- Markdown link validation failures
- ESLint errors (reduced from 2 errors, 81 warnings to 0 errors, 63 warnings)
- Improved code modularity and maintainability
