+++ src/components/common/WaitlistButton.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { joinWaitlist, leaveWaitlist, getWaitlistStatus, getWaitlistCount } from '../../services/waitlistService';
 
 // Custom hook - args grouped into single options object (fixes Excess Function Arguments)
