+++ src/components/events/LiveInventoryCounter.jsx
@@ -1,5 +1,5 @@
 
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 
 const LiveInventoryCounter = ({ eventId, initialCapacity }) => {
   const [capacity, setCapacity] = useState(initialCapacity);
