+++ src/components/ui/CopyButton.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 const CopyButton = ({ textToCopy }) => {
   const [copied, setCopied] = useState(false);
