+++ src/hooks/useOfflineSync.test.js
@@ -162,13 +162,13 @@ describe("useOfflineSync", () => {
       return null;
     };
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       root = createRoot(container);
       root.render(<TestComponent />);
     });
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       window.dispatchEvent(new Event("online"));
       await new Promise((resolve) => setTimeout(resolve, 100));
@@ -205,13 +205,13 @@ describe("useOfflineSync", () => {
       return null;
     };
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       root = createRoot(container);
       root.render(<TestComponent />);
     });
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       window.dispatchEvent(new Event("online"));
       await new Promise((resolve) => setTimeout(resolve, 100));
@@ -240,13 +240,13 @@ describe("useOfflineSync", () => {
       return null;
     };
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       root = createRoot(container);
       root.render(<TestComponent />);
     });
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       window.dispatchEvent(new Event("online"));
       await new Promise((resolve) => setTimeout(resolve, 100));
@@ -271,13 +271,13 @@ describe("useOfflineSync", () => {
       return null;
     };
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       root = createRoot(container);
       root.render(<TestComponent />);
     });
 
-    // eslint-disable-next-line testing-library/no-unnecessary-act
+     
     await act(async () => {
       window.dispatchEvent(new Event("online"));
       await new Promise((resolve) => setTimeout(resolve, 100));
